requirements.txt
Django==1.10.2
djangorestframework==3.5.2
django-redis==4.5.0
celery==3.1.21